AddOrder

+++ src/main/java/blackapple/webbook/userPart/payload/JwtResponse.java
@@ -5,12 +5,12 @@ import java.util.List;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Long id;
+    private int id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, Long id, String username, String email, List<String> roles) {
+    public JwtResponse(String accessToken, int id, String username, String email, List<String> roles) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
@@ -34,7 +34,7 @@ public class JwtResponse {
         this.type = tokenType;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -50,7 +50,7 @@ public class JwtResponse {
         return roles;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
